requirements.txt
python-rtmidi
adafruit-circuitpython-neopixel
